Rename `listener` to `listeners`
